Move tasks and deps
