+++ apps/front/package.json
@@ -14,7 +14,7 @@
     "build": "npm run build:server && npm run build:scripts && npm run build:static && npm run generate",
     "generate": "node dist/_scripts/exe-prerender.js",
     "preview": "serve dist/static",
-    "prettier": "prettier --write ",
+    "prettier": "prettier --write",
     "test:watch": "vitest",
     "test": "vitest run",
     "reset": "rm -rf dist && rm -rf node_modules && rm -rf .turbo && rm -rf build-stats.html"
