+++ apps/front/src/server/helpers/preventSlashes.ts
