+++ apps/front/tsconfig.json
@@ -23,5 +23,5 @@
       "~/*": ["./*"]
     }
   },
-  "include": ["./src"]
+  "include": ["./src", "./config"]
 }
