+++ packages/cli/src/helpers/logger.js
@@ -1,9 +1,8 @@
-import couleur from "./couleur.js"
-
-const blue = couleur.rgb(0, 180, 180)
-const grey = couleur.rgb(121, 121, 121)
-const green = couleur.rgb(57, 180, 0)
-const red = couleur.rgb(244, 34, 34)
+import chalk from "chalk"
+const blue = chalk.rgb(0, 180, 180)
+const grey = chalk.rgb(121, 121, 121)
+const green = chalk.rgb(57, 180, 0)
+const red = chalk.rgb(244, 34, 34)
 
 export default {
   start: (message = "") => console.log(`\n  ${blue(message)}`),
