--- packages/cli/src/helpers/palette.js
@@ -1,8 +1,0 @@
-import couleur from "./couleur.js"
-
-export default {
-  blue: couleur.rgb(0, 180, 180),
-  grey: couleur.rgb(121, 121, 121),
-  green: couleur.rgb(57, 180, 0),
-  red: couleur.rgb(244, 34, 34),
-}
