+++ packages/cli/src/index.js
@@ -1,6 +1,3 @@
-export { default as couleur } from "./helpers/couleur.js"
-export { default as palette } from "./helpers/palette.js"
-export { preventSlashes } from "./helpers/prevent-slashes.js"
 export { default as createFile } from "./helpers/create-file.js"
 export { default as logger } from "./helpers/logger.js"
 export * as mfs from "./helpers/mfs.js"
